Addressed wuyi's comments
